pyproject.toml
[project]
name = "auto-pa"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "anyio>=4.12.0",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "requests>=2.32.5",
    "tianxiu2b2t-utils>=0.1.36",
]
